+++ test/01-unit-initialization.js
@@ -1,1 +1,0 @@
-describe('01 - unit - initialization', function() {});