+++ test/06-unit-departing-peer.js
@@ -1,3 +1,0 @@
-describe('06 - unit - departing peer', function() {
-  it('');
-});